requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
pyrootutils>=1.0.4
pre-commit      # hooks for applying linters on commit
rich>=13.3.5
pytest>=7.3.1
# sh            # for running bash commands in some tests (linux/macos only)

omegaconf>=2.3.0
numpy>=1.24.3
opencv-python>=4.5.4.60
matplotlib>=3.7.1
kornia>=0.6.12
fastai>=2.7.12
lpips>=0.1.4
albumentations>=1.3.0
tqdm>=4.65.0
pillow>=9.5.0
pycocotools>=2.0.6
packaging>=23.1
setuptools>=66.0.0